+++ extensions/pexels/src/search-videos.tsx
@@ -1,24 +1,24 @@
 import { getPreferenceValues, Grid, List } from "@raycast/api";
 import React, { useState } from "react";
-import { searchPhotos } from "./hooks/hooks";
-import { PhotosListItem } from "./components/photos-list-item";
+import { searchVideos } from "./hooks/hooks";
 import { PexelsEmptyView } from "./components/pexels-empty-view";
 import { isEmpty } from "./utils/common-utils";
 import { SearchRequest } from "./types/types";
 import { Preferences } from "./types/preferences";
-import { PhotosGridItem } from "./components/photos-grid-item";
+import { VideosGridItem } from "./components/videos-grid-item";
+import { VideosListItem } from "./components/video-list-item";
 
-export default function SearchPhotos() {
+export default function SearchVideos() {
   const preferences = getPreferenceValues<Preferences>();
   const [searchContent, setSearchContent] = useState<string>("");
   const [searchRequest, setSearchRequest] = useState<SearchRequest>({ searchContent: searchContent, page: 1 });
-  const { pexelsPhotos, loading } = searchPhotos(searchRequest);
+  const { pexelsVideos, loading } = searchVideos(searchRequest);
 
   const emptyViewTitle = () => {
     if (loading) {
       return "Loading...";
     }
-    if (pexelsPhotos?.photos.length === 0 && !isEmpty(searchContent)) {
+    if (pexelsVideos?.videos.length === 0 && !isEmpty(searchContent)) {
       return "No Photos";
     }
     return "Welcome to Pexels";
@@ -26,7 +26,7 @@ export default function SearchPhotos() {
 
   return preferences.layout === "List" ? (
     <List
-      isShowingDetail={pexelsPhotos?.photos.length !== 0}
+      isShowingDetail={pexelsVideos?.videos.length !== 0}
       isLoading={loading}
       searchBarPlaceholder={"Search photos"}
       onSearchTextChange={(newValue) => {
@@ -35,7 +35,7 @@ export default function SearchPhotos() {
       }}
       onSelectionChange={(id) => {
         if (typeof id !== "undefined") {
-          const _id = pexelsPhotos?.photos.length - 1 + "_" + pexelsPhotos?.photos[pexelsPhotos.photos.length - 1]?.id;
+          const _id = pexelsVideos?.videos.length - 1 + "_" + pexelsVideos?.videos[pexelsVideos.videos.length - 1]?.id;
           if (id === _id) {
             setSearchRequest({ searchContent: searchContent, page: searchRequest.page + 1 });
           }
@@ -44,8 +44,8 @@ export default function SearchPhotos() {
       throttle={true}
     >
       <PexelsEmptyView title={emptyViewTitle()} layout={preferences.layout} />
-      {pexelsPhotos?.photos.map((value, index) => (
-        <PhotosListItem key={index} item={value} index={index} />
+      {pexelsVideos?.videos.map((value, index) => (
+        <VideosListItem key={index} item={value} index={index} />
       ))}
     </List>
   ) : (
@@ -61,7 +61,7 @@ export default function SearchPhotos() {
       }}
       onSelectionChange={(id) => {
         if (typeof id !== "undefined") {
-          const _id = pexelsPhotos?.photos.length - 1 + "_" + pexelsPhotos?.photos[pexelsPhotos.photos.length - 1]?.id;
+          const _id = pexelsVideos?.videos.length - 1 + "_" + pexelsVideos?.videos[pexelsVideos.videos.length - 1]?.id;
           if (id === _id) {
             setSearchRequest({ searchContent: searchContent, page: searchRequest.page + 1 });
           }
@@ -70,8 +70,8 @@ export default function SearchPhotos() {
       throttle={true}
     >
       <PexelsEmptyView title={emptyViewTitle()} layout={preferences.layout} />
-      {pexelsPhotos?.photos.map((value, index) => (
-        <PhotosGridItem key={index} item={value} index={index} />
+      {pexelsVideos?.videos.map((value, index) => (
+        <VideosGridItem key={index} item={value} index={index} />
       ))}
     </Grid>
   );