+++ extensions/pexels/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Pexels Changelog
 
+## [Add New Command] - {PR_MERGE_DATE}
+
+- Add Search Videos command
+
 ## [Update Raycast API] - 2022-12-05
 
 - Update Raycast API version to 1.44.0